Refactor scene package structure for improved organization
- Changed the package name from `ecs` to `scenes` in multiple files to better reflect their purpose within the codebase.
- Updated imports and package declarations in `gameover.go`, `manager.go`, `menu_system.go`, `menu.go`, `paused.go`, `playing.go`, `simple_text.go`, `studio_intro.go`, `title.go`, and `util.go` to maintain consistency across the scene management files.
