+++ internal/ecs/scenes/manager.go
@@ -1,4 +1,4 @@
-package ecs
+package scenes
 
 import (
 	"github.com/hajimehoshi/ebiten/v2"
