+++ internal/ecs/scenes/menu_system.go
@@ -1,4 +1,4 @@
-package ecs
+package scenes
 
 import (
 	"image/color"
