+++ internal/ecs/scenes/playing.go
@@ -1,4 +1,4 @@
-package ecs
+package scenes
 
 import (
 	"fmt"
