+++ internal/ecs/scenes/simple_text.go
@@ -1,4 +1,4 @@
-package ecs
+package scenes
 
 import (
 	"image/color"
