+++ internal/ecs/scenes/title.go
@@ -1,4 +1,4 @@
-package ecs
+package scenes
 
 import (
 	"image/color"
