+++ internal/ecs/scenes/util.go
@@ -1,4 +1,4 @@
-package ecs
+package scenes
 
 import (
 	"log"
